Initial Commit
This commit has the essential files of the website, and is the first version index.html. Inside of the index.html there is several sections which are not yet fully styled however are partially. Index.html includes a banner, map, navbar, footer, a section with bs5 and is almost fully compatible with mobile. Most of the CSS used for this section so far is with Bootstrap 5 and will be referenced appropriately in later versions.